Changes corresponding to u128 changes to Bytes

+++ src/engine/shared.rs
@@ -114,7 +114,7 @@ pub fn set_key_shared(key_fd: RawFd) -> StratisResult<SizedKeyMemory> {
                 ErrorEnum::Invalid,
                 format!(
                     "Provided key exceeded maximum allow length of {}",
-                    Bytes(MAX_STRATIS_PASS_SIZE as u64)
+                    Bytes(MAX_STRATIS_PASS_SIZE as u128)
                 ),
             ));
         }
